+++ fid-backend/fid-backend.cabal
@@ -15,15 +15,26 @@ cabal-version:       >=1.10
 
 library
   hs-source-dirs:      src
-  exposed-modules:     EthInterface
+  exposed-modules:     Docs
+                     , EthInterface
+                     , Server
   build-depends:       base >= 4.7 && < 5
                      , aeson
                      , base16-bytestring
                      , bytestring
                      , either
+                     , http-types
                      , listsafe
+                     , list-t
                      , memory
+                     , mtl
+                     , servant
+                     , servant-docs
+                     , servant-server
+                     , stm
+                     , stm-containers
                      , text
+                     , wai
                      , web3
   default-language:    Haskell2010
 
@@ -39,8 +50,6 @@ executable fid-server
                      , list-t
                      , mtl
                      , servant-server
-                     , stm
-                     , stm-containers
                      , text
                      , transformers
                      , wai
